MSE-in-Workers: Part 2: New registration and attachment logic
This CL uses a new, ref-counted, HTMLMediaSourceAttachment object as the
registrable in the MediaSourceRegistry (instead of the MediaSource
object itself), and handles the attachment and
detachment of the MSE API to an HTMLMediaElement as moderated by this
new attachment class.

Along with managing GC-traceable references in each of HTMLMediaElement
and MediaSource, HTMLMediaSourceAttachment is expected to enable, in
later CLs, cross-thread (aka MSE-in-Workers) attachments of a main
thread HTMLMediaElement to a worker thread MediaSource.

The registry singleton itself is owned by the main thread, setup by a
new Init() call executed during modules initialization. Calls to
register or unregister a MediaSource objectUrl are proxied to the main
thread if necessary. Lookups to retrieve a MediaSource by its url are
allowed only on the main thread (to satisfy the only known use case:
attachment of the MSE API to an HTMLMediaElement).

This is the second of a multi-part series of changes to enable
experimental MSE API usage from dedicated and shared worker contexts, in
addition to previous usage from the same (main) thread.

Later changes will enable instantiation of a MediaSource object on a
DedicatedWorker thread (in-process). SharedWorker support is not in this
initial feature work (deferred, see https://crbug.com/1054566).

Intent-to-implement:
https://groups.google.com/a/chromium.org/forum/#!msg/blink-dev/CNRywDqgKjY/F0nnA4tTAwAJ

BUG=878133,1054566

Change-Id: I75ca36721f239ceae643a5790c769bbf47d574bf