Don't let a CompositingInputsRoot be a contain:strict element if it was dirty.
contain:strict elements are stopping points for ancestor recursion when
finding a CompositingInputsRoot, but this should only be for when
*descendants* of the contain:strict element are dirtied, not the
contain:strict element itself.

This problem exhibited in a bug in a case when the contain:strict
element was also position:sticky, causing it to fail to update sticky
position constraints as a result.

Bug: 1051455

Change-Id: If7e9f3d19ce5dc949ee9ba6ab7c0c0c2e03b2f20